// libs/features/collection/billing/src/lib/shared/models/allocation-filter-search.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class AllocationFilterSearch {
  selectedDate: string = undefined;
  isFilter: boolean = undefined;
  isSearch: boolean = undefined;
  pageNo: number = undefined;
  pageSize: number = undefined;
  maxAllocatedAmount: number = undefined;
  maxBalanceAmount: number = undefined;
  maxDebitAmount: number = undefined;
  minAllocatedAmount: number = undefined;
  minBalanceAmount: number = undefined;
  minDebitAmount: number = undefined;
  searchKey: string = undefined;
}